+++ putils/src/main/java/io/anemos/metastore/putils/ProtoDomain.java
@@ -1,11 +1,21 @@
 package io.anemos.metastore.putils;
 
+import com.google.protobuf.Any;
+import com.google.protobuf.Api;
 import com.google.protobuf.ByteString;
 import com.google.protobuf.DescriptorProtos;
 import com.google.protobuf.Descriptors;
+import com.google.protobuf.Duration;
 import com.google.protobuf.DynamicMessage;
+import com.google.protobuf.Empty;
 import com.google.protobuf.ExtensionRegistry;
+import com.google.protobuf.FieldMask;
+import com.google.protobuf.Int32Value;
 import com.google.protobuf.InvalidProtocolBufferException;
+import com.google.protobuf.SourceContext;
+import com.google.protobuf.Struct;
+import com.google.protobuf.Timestamp;
+import com.google.protobuf.Type;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
@@ -75,13 +85,46 @@ public class ProtoDomain implements Serializable {
     }
     DescriptorProtos.FileDescriptorProto fileDescriptorProto = inMap.get(name);
     if (fileDescriptorProto == null) {
-      if ("google/protobuf/descriptor.proto".equals(name)) {
-        outMap.put(
-            "google/protobuf/descriptor.proto",
-            DescriptorProtos.FieldOptions.getDescriptor().getFile());
-        return DescriptorProtos.FieldOptions.getDescriptor().getFile();
+      Descriptors.FileDescriptor fd;
+      switch (name) {
+        case "google/protobuf/descriptor.proto":
+          fd = DescriptorProtos.FieldOptions.getDescriptor().getFile();
+          break;
+        case "google/protobuf/wrapper.proto":
+          fd = Int32Value.getDescriptor().getFile();
+          break;
+        case "google/protobuf/timestamp.proto":
+          fd = Timestamp.getDescriptor().getFile();
+          break;
+        case "google/protobuf/duration.proto":
+          fd = Duration.getDescriptor().getFile();
+          break;
+        case "google/protobuf/any.proto":
+          fd = Any.getDescriptor().getFile();
+          break;
+        case "google/protobuf/api.proto":
+          fd = Api.getDescriptor().getFile();
+          break;
+        case "google/protobuf/empty.proto":
+          fd = Empty.getDescriptor().getFile();
+          break;
+        case "google/protobuf/field_mask.proto":
+          fd = FieldMask.getDescriptor().getFile();
+          break;
+        case "google/protobuf/source_context.proto":
+          fd = SourceContext.getDescriptor().getFile();
+          break;
+        case "google/protobuf/struct.proto":
+          fd = Struct.getDescriptor().getFile();
+          break;
+        case "google/protobuf/type.proto":
+          fd = Type.getDescriptor().getFile();
+          break;
+        default:
+          return null;
       }
-      return null;
+      outMap.put(name, fd);
+      return fd;
     } else {
       List<Descriptors.FileDescriptor> dependencies = new ArrayList<>();
       if (fileDescriptorProto.getDependencyCount() > 0) {